strip_tags importable as library
