+++ strip_tags/__init__.py
@@ -1,0 +1,3 @@
+from .lib import strip_tags
+
+__all__ = ["strip_tags"]
